+++ src/gui/widgets/folder_channel.c
@@ -74,7 +74,8 @@ folder_channel_snapshot (
         &Z_GDK_RGBA_INIT (
           track->color.red, track->color.green,
           track->color.blue, 0.15f),
-        &GRAPHENE_RECT_INIT (0, 0, width, height));
+        &GRAPHENE_RECT_INIT (
+          0.f, 0.f, (float) width, (float) height));
     }
 
   GTK_WIDGET_CLASS (folder_channel_widget_parent_class)
